+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - nodejs: delete file without reloading the page
 - nodejs: getting history by a separate request
 - nodejs: restore from history
+- python: restore from history
 
 ## 1.6.0
 - nodejs: setUsers for region protection